@@ -3,8 +3,12 @@ version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
 
-java.targetCompatibility = JavaVersion.VERSION_11
-java.sourceCompatibility = JavaVersion.VERSION_11
+java.targetCompatibility = JavaVersion.VERSION_19
+java.sourceCompatibility = JavaVersion.VERSION_19
+
+compileJava {
+    options.compilerArgs << '--enable-preview'
+}
 
 repositories {
     mavenCentral()
