@@ -1,14 +1,16 @@
+import common.Options;
 import tests.CompletableFuturesTests;
 import tests.ParallelStreamsTests;
 import tests.StructuredConcurrencyTests;
 import utils.*;
 
 import java.util.concurrent.*;
+import java.util.function.Consumer;
 
 /**
  * This example compares and contrasts the programming models and
  * performance results of Java parallel streams, completable futures,
- * and Project Loom structured concurrency when applied to download,
+ * and Java structured concurrency when applied to download,
  * transform, and store many images from a remote web server.
  */
 public class ex3 {
@@ -24,36 +26,39 @@ public class ex3 {
         // Initializes the Options singleton.
         Options.instance().parseArgs(argv);
 
+        runTest(ParallelStreamsTests::run,
+                "warmup common fork-join pool");
+
         // Runs the tests using the Java parallel streams framework.
         runTest(ParallelStreamsTests::run,
-                "Parallel streams run() test");
+                "Parallel streams implementation");
 
         // Runs the tests using the Java completable futures framework.
         runTest(CompletableFuturesTests::run,
-                "Completable futures runFineGrained() test");
+                "Completable futures implementation");
 
         // Runs the tests using the Java framework.
         runTest(StructuredConcurrencyTests::run,
-                "Structured concurrency runFineGrained() test");
+                "Structured concurrency implementation");
 
         System.out.println(RunTimer.getTimingResults());
     }
 
     /**
-     * Record the amount of time needed to runFineGrained test.
+     * Record the amount of time needed to run {@code runTest}.
      */
-    private static void runTest(Runnable runTest,
+    private static void runTest(Consumer<String> test,
                                 String testName) {
-        // Let the system garbage collect.
+        // Let the system garbage collect first to ensure pristine conditions.
         System.gc();
 
-        // Record how long the test takes to runFineGrained.
+        // Record how long the test takes to run runTest.
         // Run the test with the designated
         // functions.
-        RunTimer.timeRun(runTest,
+        RunTimer.timeRun(() -> test.accept(testName),
                          testName);
 
-        // Delete any images from the previous runFineGrained.
+        // Delete any images from the previous run.
         FileAndNetUtils
             .deleteDownloadedImages(Options.instance().getDirectoryPath());
     }
